Added global function to send requests (for tests).
