+++ generators/app/templates/test/unit/controllers/AuthController.test.js
@@ -1,23 +1,15 @@
 var assert = require('assert');
 var AuthController = require('../../../api/controllers/AuthController');
 var Promise = require('bluebird');
-var supertest = require('supertest');
 
 var Users = require('../../fixtures/Users.json');
 var FaultyUsers = require('../../fixtures/FaultyUsers.json');
-
-function request(method, url) {
-  return supertest(sails.hooks.http.app)[method](url)
-    .set('Content-Type', 'application/json')
-    .set('Accept', 'application/json')
-    .expect('Content-Type', /json/);
-}
 
 describe("controllers:AuthController", function () {
   it("should register new users", function (done) {
     Promise.map(Users, function (user) {
       return new Promise(function (resolve, reject) {
-        request('post', '/v1/auth/signup')
+        sails.requestForTest('post', '/v1/auth/signup')
           .send(user)
           .expect(201)
           .end(function (err, data) {
@@ -36,7 +28,7 @@ describe("controllers:AuthController", function () {
   it("should login with just created users", function (done) {
     Promise.map(Users, function (user) {
       return new Promise(function (resolve, reject) {
-        request('post', '/v1/auth/signin')
+        sails.requestForTest('post', '/v1/auth/signin')
           .send(user)
           .expect(200)
           .end(function (err, data) {
@@ -60,7 +52,7 @@ describe("controllers:AuthController", function () {
   it("should be rejected after trying to signup with faulty users", function (done) {
     Promise.map(FaultyUsers, function (user) {
       return new Promise(function (resolve, reject) {
-        request('post', '/v1/auth/signup')
+        sails.requestForTest('post', '/v1/auth/signup')
           .send(user)
           .end(function (err, data) {
             if (err) return resolve(err);
